+++ src/libHeaven/CentralUI/States/WindowStateRoot.hpp
@@ -16,25 +16,34 @@
  *
  */
 
-#ifndef MGV_HEAVEN_MODE_PRIVATE_H
-#define MGV_HEAVEN_MODE_PRIVATE_H
+#ifndef HEAVEN_WINDOW_STATE_ROOT_HPP
+#define HEAVEN_WINDOW_STATE_ROOT_HPP
 
-#include <QString>
-
-#include "CentralUI/States/WindowStateRoot.hpp"
+#include "libHeaven/CentralUI/States/WindowState.hpp"
 
 namespace Heaven
 {
 
-    class ModePrivate
+    class WindowStateRoot : public WindowState
     {
     public:
-        QString                 mName;
-        WindowStateRoot::Ptr    mRoot;
-        WindowStateRoot::Ptr    mOriginalState;
+        typedef QExplicitlySharedDataPointer< WindowStateRoot > Ptr;
+
+    public:
+        WindowStateRoot();
+        WindowStateRoot( const QDomElement& elParent );
+
+        ~WindowStateRoot();
+
+    public:
+        Type type() const;
+
+        void save( QDomElement& elParent ) const;
+
+    private:
+        void load( const QDomElement& elParent );
     };
 
 }
-
 
 #endif
