+++ src/libHeaven/CentralUI/States/WindowStateView.hpp
@@ -16,25 +16,32 @@
  *
  */
 
-#ifndef MGV_HEAVEN_MODE_PRIVATE_H
-#define MGV_HEAVEN_MODE_PRIVATE_H
+#ifndef HEAVEN_WINDOW_STATE_VIEW_HPP
+#define HEAVEN_WINDOW_STATE_VIEW_HPP
 
-#include <QString>
-
-#include "CentralUI/States/WindowStateRoot.hpp"
+#include "libHeaven/CentralUI/States/WindowState.hpp"
 
 namespace Heaven
 {
 
-    class ModePrivate
+    class WindowStateView : public WindowState
     {
     public:
-        QString                 mName;
-        WindowStateRoot::Ptr    mRoot;
-        WindowStateRoot::Ptr    mOriginalState;
+        typedef QExplicitlySharedDataPointer< WindowStateView > Ptr;
+
+    public:
+        WindowStateView( WindowState* parent );
+        WindowStateView( WindowState* parent, QDomElement& el );
+
+    public:
+        Type type() const;
+
+        void updateConfig();
+
+    protected:
+        void save( QDomElement& elParent ) const;
     };
 
 }
-
 
 #endif
