+++ src/libHeaven/CentralUI/States/WindowStateWindow.hpp
@@ -16,25 +16,30 @@
  *
  */
 
-#ifndef MGV_HEAVEN_MODE_PRIVATE_H
-#define MGV_HEAVEN_MODE_PRIVATE_H
+#ifndef HEAVEN_WINDOW_STATE_WINDOW_HPP
+#define HEAVEN_WINDOW_STATE_WINDOW_HPP
 
-#include <QString>
-
-#include "CentralUI/States/WindowStateRoot.hpp"
+#include "libHeaven/CentralUI/States/WindowState.hpp"
 
 namespace Heaven
 {
 
-    class ModePrivate
+    class WindowStateWindow : public WindowState
     {
     public:
-        QString                 mName;
-        WindowStateRoot::Ptr    mRoot;
-        WindowStateRoot::Ptr    mOriginalState;
+        typedef QExplicitlySharedDataPointer< WindowStateWindow > Ptr;
+
+    public:
+        WindowStateWindow( WindowState* parent );
+        WindowStateWindow( WindowState* parent, QDomElement& el );
+
+    public:
+        Type type() const;
+
+    protected:
+        void save( QDomElement& elParent ) const;
     };
 
 }
-
 
 #endif
